move package into model_compression_toolkit
