+++ model_compression_toolkit/xquant/common/model_folding_utils.py
