+++ model_compression_toolkit/xquant/pytorch/model_analyzer_utils.py
@@ -17,7 +17,7 @@ from typing import Dict, List, Tuple
 import torch
 from mct_quantizers.pytorch.quantize_wrapper import PytorchQuantizationWrapper
 
-from xquant.common.model_analyzer_utils import ModelAnalyzerUtils
+from model_compression_toolkit.xquant.common.model_analyzer_utils import ModelAnalyzerUtils
 
 
 class PytorchModelAnalyzerUtils(ModelAnalyzerUtils):
